CertificateSubmisisonStrategy added: ceasing and non-ceasing.
