+++ sdk/src/main/scala/com/horizen/certificatesubmitter/network/CertificateSignaturesMessages.scala
@@ -1,6 +1,6 @@
 package com.horizen.certificatesubmitter.network
 
-import com.horizen.certificatesubmitter.AbstractCertificateSubmitter.CertificateSignatureInfo
+import com.horizen.certificatesubmitter.CertificateSubmitter.CertificateSignatureInfo
 import com.horizen.cryptolibprovider.FieldElementUtils
 import com.horizen.proof.SchnorrSignatureSerializer
 import sparkz.core.network.message.{Message, MessageSpecV1}
