+++ sdk/src/main/scala/com/horizen/certificatesubmitter/strategies/CertificateSubmissionStrategy.scala
@@ -1,0 +1,14 @@
+package com.horizen.certificatesubmitter.strategies
+
+import com.horizen.block.SidechainBlock
+import com.horizen.{SidechainHistory, SidechainMemoryPool, SidechainState, SidechainWallet}
+import com.horizen.certificatesubmitter.CertificateSubmitter.SignaturesStatus
+import sparkz.core.NodeViewHolder.CurrentView
+
+trait CertificateSubmissionStrategy {
+  type View = CurrentView[SidechainHistory, SidechainState, SidechainWallet, SidechainMemoryPool]
+
+  def getStatus(sidechainNodeView: View, block: SidechainBlock): SubmissionWindowStatus
+  def checkQuality(status: SignaturesStatus): Boolean
+
+}
