+++ sdk/src/main/scala/com/horizen/certificatesubmitter/strategies/SubmissionWindowStatus.scala
@@ -1,0 +1,3 @@
+package com.horizen.certificatesubmitter.strategies
+
+case class SubmissionWindowStatus(referencedWithdrawalEpochNumber: Int, isInWindow: Boolean)
