+++ sdk/src/test/scala/com/horizen/certificatesubmitter/network/CertificateSignaturesManagerTest.scala
@@ -3,8 +3,8 @@ package com.horizen.certificatesubmitter.network
 import akka.actor.{ActorRef, ActorSystem, Props}
 import akka.testkit.{TestActor, TestActorRef, TestProbe}
 import akka.util.Timeout
-import com.horizen.certificatesubmitter.AbstractCertificateSubmitter.ReceivableMessages.{GetSignaturesStatus, SignatureFromRemote}
-import com.horizen.certificatesubmitter.AbstractCertificateSubmitter.{BroadcastLocallyGeneratedSignature, CertificateSignatureFromRemoteInfo, CertificateSignatureInfo, DifferentMessageToSign, InvalidPublicKeyIndex, InvalidSignature, KnownSignature, SignatureProcessingStatus, SignaturesStatus, SubmitterIsOutsideSubmissionWindow, ValidSignature}
+import com.horizen.certificatesubmitter.CertificateSubmitter.ReceivableMessages.{GetSignaturesStatus, SignatureFromRemote}
+import com.horizen.certificatesubmitter.CertificateSubmitter.{BroadcastLocallyGeneratedSignature, CertificateSignatureFromRemoteInfo, CertificateSignatureInfo, DifferentMessageToSign, InvalidPublicKeyIndex, InvalidSignature, KnownSignature, SignatureProcessingStatus, SignaturesStatus, SubmitterIsOutsideSubmissionWindow, ValidSignature}
 import com.horizen.certificatesubmitter.network.CertificateSignaturesManager.InternalReceivableMessages.TryToSendGetCertificateSignatures
 import com.horizen.fixtures.FieldElementFixture
 import com.horizen.SidechainAppEvents
@@ -26,7 +26,7 @@ import scala.collection.mutable.ArrayBuffer
 import scala.concurrent.ExecutionContext
 import scala.language.postfixOps
 import scala.concurrent.duration._
-import scala.util.Success
+import scala.util.{Failure, Success}
 
 class CertificateSignaturesManagerTest extends JUnitSuite with MockitoSugar {
 
@@ -44,6 +44,8 @@ class CertificateSignaturesManagerTest extends JUnitSuite with MockitoSugar {
   def roundTrip(msg: Message[_]): Message[_] = {
     messageSerializer.deserialize(messageSerializer.serialize(msg), msg.source) match {
       case Success(Some(value)) => value
+      case Success(_) => fail("unexpected case")
+      case Failure(ex) => fail("round trip serialization failed", ex)
     }
   }
